pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cost-guard"
version = "0.1.0"
description = "Financial safety tool for AI API costs"
authors = [{name = "AI Cost Guard", email = "dev@aicostguard.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "typer>=0.9.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
ai-cost-guard = "ai_cost_guard.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_cost_guard*"]